+++ docs/conf.py
@@ -13,7 +13,7 @@
 import os
 import sys
 sys.path.append(os.path.abspath(".."))
-sys.path.append(os.path.abspath("../pluralkit/v1"))
+sys.path.append(os.path.abspath("../pluralkit/v2"))
 sys.path.insert(0, os.path.abspath(".."))
 from pluralkit import __version__
 from datetime import datetime
@@ -21,7 +21,7 @@ from datetime import datetime
 # -- Project information -----------------------------------------------------
 
 project = "pluralkit.py"
-copyright = "2021, Madison Landry, Alyx Warner"
+copyright = "2022, Madison Landry, Alyx Warner"
 author = "Madison Landry, Alyx Warner"
 
 # The full version